Create file todolist.js to interactive functions.
